Add files via upload
Chroniclers Modified Rolls is a module that I originally created for my own private group. It allows the GM to modify a roll in chat after the result has been posted. This simply happens by simply right-clicking on the roll message, which opens up the context menu with the new option "Modify this roll". This option then allows the GM to choose a number (positive or negative) that will modify the result of the roll accordingly. An italic text "Modifier" will be appended to the flavor text of the roll to indicate the change.

This option should only appear for the GM and only if the chosen message is a roll.